This is a 2 player game designed in Python using its GUI features.
So the game is like this:
The goal of the players is to keep the ball within the court. The player
who fails to do that looses a point to the opposite player. As the time
passes, the speed of the ball increases. Game can be restarted at any
time by clicking the "restart" button.
Control for player1 are "w" for up and "s" for down while for player2
the controls are "up arrow" for up and "down arrow" for down.